Extract 'syntax' tests.
